init pass at reindexer singleton
